// frontend/src/queries/config/graphql-generated-types.d.ts
export type GetProjectsQuery = { __typename?: 'query_root', projects: Array<{ __typename?: 'projects', name: string, uuid: any, languages: Array<{ __typename?: 'project_language_link', uuid: any, language: { __typename?: 'languages', iso_code: string, uuid: any } }>, company: { __typename?: 'companies', name: string, uuid: any } }> };